got most of it done, still need to port the text, commands, and ios over to new contacts scheme
