+++ gui/qt/cashacctqt.py
@@ -472,7 +472,7 @@ def lookup_cash_account_dialog(
     d.setObjectName("WindowModalDialog - " + title)
     finalization_print_error(d)
     destroyed_print_error(d)
-    all_cashacct_contacts = set(v[1] for k,v in wallet.contacts.items() if v[0] == 'cashacct' )
+    all_cashacct_contacts = set(contact for contact in wallet.contacts.get_all(nocopy=True) if contact.type == 'cashacct')
 
     vbox = QVBoxLayout(d)
     hbox = QHBoxLayout()
